requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.2.5
idna==3.7
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.21.6
Pillow==9.5.0
requests==2.31.0
torch==1.13.1
torchvision==0.14.1
typing-extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
